+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-05-14 11:46:34
-// Commit: feat(state): add bi-directional event bus handler for state
+// Updated at: 2026-05-14 12:13:12
+// Commit: fix(state): fix memory leak by cancelling state subscription
 
 class stateComponentState {
-  final String commitRef = '1133';
-  final String lastMessage = 'feat(state): add bi-directional event bus handler for state';
-  final DateTime timestamp = DateTime.parse('2026-05-14T11:46:34+05:00');
+  final String commitRef = '1135';
+  final String lastMessage = 'fix(state): fix memory leak by cancelling state subscription';
+  final DateTime timestamp = DateTime.parse('2026-05-14T12:13:12+05:00');
 }